fix: address PR #29 review comments
- Add idempotentHint to update_assignment and delete_assignment annotations
- Move CreateAssignmentParams and UpdateAssignmentParams to src/canvas/types.ts
- Derive UpdateAssignmentParams as Partial<CreateAssignmentParams>
- Update test expectations for new idempotentHint annotations
